앱 수정

+++ AndroidApp/app/src/main/java/kr/ac/kumoh/ce/s20190348/AndroidAppProject/Sensor.kt
@@ -3,6 +3,6 @@ package kr.ac.kumoh.ce.s20190348.AndroidAppProject
 data class Sensor(
     val temperature : Int,
     val water : Int,
-    val gas : Float,
+    val gas : Int,
     val nfc : String,
 )
